Move ClientLoader to ktor-client-test-base
