+++ gradle/compatibility.gradle
@@ -1,8 +1,13 @@
+/*
+ * Copyright 2014-2025 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
 apply plugin: 'binary-compatibility-validator'
 
 apiValidation {
     def excludeList = [
         'ktor',
+        "ktor-client-test-base",
         'ktor-client-tests',
         'ktor-client-js',
         'ktor-client-content-negotiation-tests',
