+++ ktor-client/ktor-client-jetty-jakarta/jvm/test/io/ktor/client/engine/jetty/jakarta/JettyHttp2EngineTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014-2023 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ * Copyright 2014-2025 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
  */
 
 package io.ktor.client.engine.jetty.jakarta
@@ -10,9 +10,10 @@ import io.ktor.client.plugins.*
 import io.ktor.client.request.*
 import io.ktor.client.tests.utils.*
 import io.ktor.test.dispatcher.*
-import io.ktor.utils.io.errors.*
 import kotlinx.io.IOException
-import kotlin.test.*
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertFailsWith
 
 class JettyHttp2EngineTest {
