+++ ktor-client/ktor-client-plugins/ktor-client-content-negotiation/ktor-client-content-negotiation-tests/jvm/src/io/ktor/client/plugins/contentnegotiation/tests/AbstractClientContentNegotiationTest.kt
@@ -1,10 +1,10 @@
 /*
- * Copyright 2014-2024 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ * Copyright 2014-2025 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
  */
 
 package io.ktor.client.plugins.contentnegotiation.tests
 
-import com.fasterxml.jackson.annotation.*
+import com.fasterxml.jackson.annotation.JsonTypeInfo
 import io.ktor.client.call.*
 import io.ktor.client.engine.cio.*
 import io.ktor.client.engine.mock.*
@@ -26,7 +26,7 @@ import io.ktor.server.websocket.WebSockets
 import io.ktor.utils.io.*
 import io.ktor.websocket.*
 import kotlinx.serialization.*
-import kotlinx.serialization.builtins.*
+import kotlinx.serialization.builtins.ListSerializer
 import kotlin.test.*
 
 /** Base class for [ContentNegotiation] tests. */
@@ -176,7 +176,7 @@ abstract class AbstractClientContentNegotiationTest : TestWithKtor() {
         )
 
         test { client ->
-            val cause = kotlin.test.assertFailsWith<JsonConvertException> {
+            val cause = assertFailsWith<JsonConvertException> {
                 client.post {
                     setBody(widget)
                     url(path = "/widget", port = serverPort)
