+++ ktor-client/ktor-client-test-base/build.gradle.kts
@@ -1,0 +1,20 @@
+/*
+ * Copyright 2014-2025 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+description = ""
+
+kotlin.sourceSets {
+    commonMain {
+        dependencies {
+            api(project(":ktor-client:ktor-client-core"))
+            api(project(":ktor-shared:ktor-test-base"))
+        }
+    }
+
+    jvmMain {
+        dependencies {
+            implementation(libs.kotlinx.coroutines.debug)
+        }
+    }
+}
