+++ ktor-client/ktor-client-test-base/common/src/io/ktor/client/tests/utils/TestInfo.kt
@@ -1,0 +1,7 @@
+/*
+ * Copyright 2014-2025 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package io.ktor.client.tests.utils
+
+class TestInfo(val threadId: Int, val attempt: Int)
