+++ ktor-client/ktor-client-test-base/nonJvm/src/io/ktor/client/tests/utils/ClientLoader.nonJvm.kt
@@ -1,0 +1,9 @@
+/*
+ * Copyright 2014-2025 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package io.ktor.client.tests.utils
+
+// supported only on JVM
+internal actual fun platformDumpCoroutines() {}
+internal actual fun platformWaitForAllCoroutines() {}
