--- ktor-client/ktor-client-tests/nonJvm/src/io/ktor/client/tests/utils/ClientLoader.nonJvm.kt
@@ -1,9 +1,0 @@
-/*
- * Copyright 2014-2019 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
- */
-
-package io.ktor.client.tests.utils
-
-// supported only on JVM
-internal actual fun platformDumpCoroutines() {}
-internal actual fun platformWaitForAllCoroutines() {}
