+++ settings.gradle.kts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014-2024 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ * Copyright 2014-2025 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
  */
 
 pluginManagement {
@@ -36,6 +36,7 @@ include(":ktor-server:ktor-server-netty")
 include(":ktor-server:ktor-server-cio")
 include(":ktor-client")
 include(":ktor-client:ktor-client-core")
+include(":ktor-client:ktor-client-test-base")
 include(":ktor-client:ktor-client-tests")
 include(":ktor-client:ktor-client-apache")
 include(":ktor-client:ktor-client-apache5")
